selinux: cache access vector decisions in the inode security blob
I think Linus suggested this a long long time ago but I never got around
to trying it. Better late than never. Compute the access vector decision
when the inode security blob is initialized and cache it in the blob.
Update it on file opens. In selinux_inode_permission and inode_has_perm,
use this cached decision unless invalidated.

Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>