selinux: re-initialize the inode security blob for correct namespace
Rather than always using the current selinux namespace, pass the selinux
namespace explicitly down to inode_doinit_with_dentry() for use in
initializing or refreshing the inode security blob. This is particularly
necessary when updating the security blob for a new SELinux namespace.

If the inode security blob has already been initialized but not in the
current namespace and has an initial SID other than the unlabeled SID,
leave the initial SID intact since it remains valid under any policy.

If re-initializing for a different namespace and not using xattrs,
then convert the incore inode SID to a context in the old namespace
and then convert the context to a SID in the new namespace.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>