selinux: use the incore inode SID if not unlabeled
When re-initializing the inode SID for a different SELinux namespace,
for any SID other than the unlabeled SID, try converting the SID to a
context in the old namespace and then converting the context to a SID
in the new namespace. Previously, this was only done if not using xattrs.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>