selinux: init inode from nearest initialized namespace
Previously inode_doinit_with_dentry() was only checking sbsec->flags
to see if it should defer the inode security blob initialization,
which was fine when there was only a single SELinux state/namespace
since that could only be set if the state was initialized. However,
with the introduction of SELinux namespaces, the superblock could be
initialized in the parent (or other ancestor) namespace but the
current namespace may not yet be initialized (i.e. the namespace was
unshared but no policy has yet been loaded into it). This would
currently result in leaving the inode with the unlabeled SID and
marking it initialized.

Explicitly check whether the current selinux state is initialized, and
if not, try to find the nearest ancestor namespace that is initialized
and use it. If no initialized state can be found, defer initialization
of the inode until later.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>